Creating mapping for CSV file to database
After plenty of research, it will be easier to implement my own mapping system. The imported file is saved into its own directory and will be read from a different page to map the different data into the database. This will be done by sending the column names to the associated table headers of the CSV file through a form and then matching the data to these columns.