Import update and Job Ticket Fixes
When Importing a CSV file into the sales table, it will automatically enter a date_added field if there isn't something already selected. There is also an information box which explains basic syntax of input. In the job ticket, weights and measures can now be edited in both the regular job ticket and archive.