Import error messages and error code added
The importer now gives the association of the column header to the database header, ensuring a more specific error message and the ability to pinpoint errors in the CSV file. Also different error checks were created to make sure input doesn't exceed 45 characters, full_name, prefix, and suffix don't contain any numerical characters, and address_line_3 can be used only if country includes a field (not including the United States). New guidelines are shown on the Upload.php page which give specific instructions on the proper inputs allowed in the importing process. As a result, the database has been modified with a country field.