Minor fixes and importer updates
Fixed some styling issues on the sidebar. The yellow sheet now checks if all 14 boxes are marked as N/A meaning the job ticket can now be transferred to a different department. A minor bug was fixed in the calendar app where it wasn't able to show current date and it's agenda. In the importer, There is an import name field to name the import created. Two columns were added to the sales table called import_name and import_status. Import_name corresponds to the name of the import and import_status will be applied to a view page which will show clients inserted and updated as a result of the import.