Inventory Update
Using the vendor "CRST Inventory", whenever a user uses this as a weights and measure, it will check the inventory to see if there are enough materials. When a job ticket is created using the Cornerstone materials, it will automatically decrease from the total. Eventually, there will be a table which will give any updates when there isn't enough inventory to complete the current job tickets. Once the user changes the stock of the material, this alert will go away.