Tests and Maximum CSV set
After many tests with the importer, a maximum has been set for the CSV file because of the amount of time it takes to import the file. A CSV file can contain only 200 clients per download to save time and also to restrict the possibility of a massive CSV file being uploaded and crashing the website.