New Stylings and Bug Fixes
The date styling on each documentation page resembles a blog post and the profile pictures are now included in the documentation block for the user who created the post. A default picture is placed if the user does not have a profile picture. Also, there was a problem displaying the text area content. This has been resolved.